innit completed

+++ Habit-Tracker/src/components/Habit/Habit.tsx
@@ -4,6 +4,7 @@ import { useNavigate, useParams } from "react-router-dom";
 import Menu from "./../home/Menu";
 import Stage from "./Stage";
 import Card from "./Card";
+import Instructions from "../home/Instructions";
 
 interface Users {
   userID: string;
@@ -30,7 +31,7 @@ const Habit = () => {
   const [habit, setHabit] = useState<Habit>({
     userID: "",
     name: "",
-    numberOfDays: -1,
+    numberOfDays: 0,
     days: [],
     currDay: -1,
   });
@@ -40,13 +41,57 @@ const Habit = () => {
   const [menuOpen, setMenuOpen] = useState(false);
   const { habitID } = useParams<{ habitID: string }>();
   const { userID } = useParams<{ userID: string }>();
-  const [tasks, setTasks] = useState<string[]>();
+  const [tasks, setTasks] = useState<string[]>([]);
   const [note, setNote] = useState<string>();
-  const [date, setDate] = useState<Date>();
+  const [date, setDate] = useState<Date>(new Date());
   const [thisDay, setThisDay] = useState<number>(-1);
   const [isCardVisibile, setCardVisibility] = useState<boolean>(false);
 
   const navigate = useNavigate();
+
+  const instruction = (
+    <>
+      <h5>⚙️ How It Works</h5>
+      <p>
+        Each day is a stage you must complete to move forward. The goal is to
+        progress step by step by accomplishing tasks and maintaining
+        consistency.
+      </p>
+      <p>
+        Schedule your tasks for tomorrow and complete them to unlock the next
+        stage. Your progress is your power!
+      </p>
+
+      <h5>📊 How to Track Your Progress</h5>
+      <p>
+        - Click on the next stage (marked in white) and select <b>Load</b> to
+        retrieve any previously scheduled tasks.
+      </p>
+      <p>
+        - Plan your tasks for tomorrow by flipping the card (click <b>{`<`}</b>
+        ), add any notes on the backside, and then click <b>Save</b> to confirm
+        your tasks.
+      </p>
+      <p>
+        - After saving, click <b>Are you doing all these tomorrow?</b> to
+        finalize your schedule.
+      </p>
+
+      <h5>✅ Completing Tasks and Moving Ahead</h5>
+      <p>
+        On the next day, complete your scheduled tasks and mark them as{" "}
+        <b>Completed</b>. Once done, move to the next stage and schedule your
+        tasks for the following day.
+      </p>
+      <p>Keep going, and watch your habits transform into a routine!</p>
+      <h5>🗑️ Deleting a Habit</h5>
+      <p>
+        To remove a habit, open the menu by clicking on the menu icon and select
+        <b>Delete Habit</b>. Confirm the deletion, and the habit will be removed
+        from your list.
+      </p>
+    </>
+  );
 
   useEffect(() => {
     try {
@@ -59,6 +104,24 @@ const Habit = () => {
       const userDetails = axios.get(
         `${import.meta.env.VITE_BASE_API_URL_V1}/${userID}/user-details`
       );
+      const habitDetails = axios.get(
+        `${import.meta.env.VITE_BASE_API_URL_V1}/${userID}/habits/${habitID}`
+      );
+      habitDetails
+        .then((resp) => {
+          if (resp.status == 200) {
+            return resp.data;
+          }
+        })
+        .then((data: Habit) => {
+          setLoading(false);
+          setHabit(data);
+        })
+        .catch(() => {
+          setLoading(false);
+          setError(true);
+          navigate("/login");
+        });
       userDetails
         .then((resp) => {
           if (resp.status == 200) {
@@ -80,36 +143,13 @@ const Habit = () => {
 
       navigate("/login");
     }
-  }, [navigate, userID]);
-  useEffect(() => {
-    try {
-      const userDetails = axios.get(
-        `${import.meta.env.VITE_BASE_API_URL_V1}/${userID}/habits/${habitID}`
-      );
-      userDetails
-        .then((resp) => {
-          if (resp.status == 200) {
-            return resp.data;
-          }
-        })
-        .then((data: Habit) => {
-          setLoading(false);
-          setHabit(data);
-        })
-        .catch(() => {
-          setLoading(false);
-          setError(true);
-          navigate("/login");
-        });
-    } catch {
-      setLoading(false);
-      setError(true);
-
-      navigate("/login");
-    }
-  }, [navigate, habitID, userID]);
+  }, [navigate, userID, habitID]);
 
   const deleteHabit = async () => {
+    const confirmDelete = confirm(
+      "Are you sure? \nNote: All your progress will be deleted."
+    );
+    if (!confirmDelete) return;
     try {
       const result = await axios.delete(
         `${import.meta.env.VITE_BASE_API_URL_V1}/${userID}/habits/${habitID}`
@@ -128,88 +168,87 @@ const Habit = () => {
   const onStageSelect = (day: number) => {
     setCardVisibility(true);
     setThisDay(day);
+
+    setTasks(habit?.days[day] ? habit?.days[day].tasks : []);
+    setNote(habit?.days[day] ? habit?.days[day].note : "");
     if (day == habit?.currDay) {
-      setTasks([]);
-      setNote(undefined);
-      setDate(new Date());
-      return;
+      setDate(new Date(new Date()));
+      date.setDate(new Date().getDate() + 1);
+      setDate(date);
+    } else {
+      setDate(new Date(habit?.days[day].date));
     }
-    setTasks(habit?.days[day].tasks);
-    setNote(habit?.days[day].note);
-    setDate(habit?.days[day].date);
   };
 
   if (loading) return <h1>LOADING</h1>;
   if (error) return <h1>ERROR</h1>;
 
   return (
-    <div
-      style={{ display: isCardVisibile ? "block" : "none" }}
-      className="d-flex vh-100 position-relative"
-    >
-      {/* Sidebar */}
-      <div className="row">
-        <div className="col">
-          <Menu
-            menuOpen={menuOpen}
-            setMenuOpen={setMenuOpen}
-            heading={userdata?.userID}
-            options={["Home", "Delete Habit"]}
-            onClicks={[
-              () => {
-                navigate(`/${userID}`);
-              },
-              deleteHabit,
-            ]}
-          ></Menu>
+    <div>
+      {message && (
+        <div className="alert alert-light text-center mb-auto" role="alert">
+          {message}
         </div>
-        <div className="col">
-          {message && (
-            <div className="alert alert-light text-center mb-auto" role="alert">
-              {message}
-            </div>
-          )}
-        </div>
-        <div className="col">
-          <div className="card p-3">Day={habit?.currDay}</div>
-        </div>
+      )}
+      <div className="d-flex justify-content-end m-3">
+        <Instructions title={habit.name}>{instruction}</Instructions>
       </div>
-
-      {/* Main Content */}
-
       <div
-        className={`flex-grow-1 ms-auto ${menuOpen ? "content-shifted" : ""}`}
-        style={{ overflowX: "auto", whiteSpace: "nowrap", padding: "20px" }}
+        style={{ display: isCardVisibile ? "block" : "none" }}
+        className="d-flex vh-100 position-relative"
       >
-        <div className="d-flex justify-content-center shadow">
-          <Card
-            tasks={tasks}
-            setTasks={setTasks}
-            note={note}
-            setNote={setNote}
-            date={date}
-            setDate={setDate}
-            isVisibile={isCardVisibile}
-            setVisibility={setCardVisibility}
-            thisDay={thisDay}
-            currDay={habit.currDay}
-          ></Card>
+        {/* Sidebar */}
+        <div className="row">
+          <div className="col">
+            <Menu
+              menuOpen={menuOpen}
+              setMenuOpen={setMenuOpen}
+              heading={userdata?.userID}
+              options={["Home", "Delete Habit"]}
+              onClicks={[
+                () => {
+                  navigate(`/${userID}`);
+                },
+                deleteHabit,
+              ]}
+            ></Menu>
+          </div>
+
+          <div className="col">
+            <div className="card p-3">Day={habit?.currDay}</div>
+          </div>
         </div>
-        {/* Scrollable Content */}
-        <div className="d-inline-flex mt-5 d-flex align-items-center vh-100">
-          {[...Array(habit?.numberOfDays)].map((_, i) => (
-            <div
-              key={i}
-              onClick={() => {
-                onStageSelect(i);
-              }}
-            >
-              <Stage
-                num={i <= habit?.currDay ? `${i + 1}` : "🔒"}
-                currDay={habit?.currDay}
-              ></Stage>
-            </div>
-          ))}
+        {/* Main Content */}
+        <div
+          className={`flex-grow-1 ms-auto ${menuOpen ? "content-shifted" : ""}`}
+          style={{ overflowX: "auto", whiteSpace: "nowrap", padding: "20px" }}
+        >
+          <div className="d-flex justify-content-center shadow">
+            <Card
+              tasks={tasks}
+              setTasks={setTasks}
+              note={note}
+              setNote={setNote}
+              date={date}
+              isVisibile={isCardVisibile}
+              setVisibility={setCardVisibility}
+              thisDay={thisDay}
+              currDay={habit.currDay}
+            ></Card>
+          </div>
+          {/* Scrollable Content */}
+          <div className="d-inline-flex mt-5 d-flex align-items-center vh-100">
+            {[...Array(habit?.numberOfDays)].map((_, i) => (
+              <div key={i}>
+                <Stage
+                  num={i <= habit?.currDay ? `${i + 1}` : "🔒"}
+                  day={i}
+                  currDay={habit?.currDay}
+                  onStageSelect={onStageSelect}
+                ></Stage>
+              </div>
+            ))}
+          </div>
         </div>
       </div>
     </div>
